refactor: move workflow variable replacement to service
