+++ packages/global/common/string/tools.ts
@@ -1,19 +1,6 @@
 import crypto from 'crypto';
 import { customAlphabet } from 'nanoid';
 import path from 'path';
-import { getErrText } from '../error/utils';
-
-export const DEFAULT_MAX_STRING_LENGTH = 100_000_000;
-
-export const getTextOversizeErrorMessage = (size = DEFAULT_MAX_STRING_LENGTH) =>
-  `Text length exceeds ${size.toLocaleString('en-US')} characters.`;
-
-export const checkStrOversize = (str: string, size = DEFAULT_MAX_STRING_LENGTH) => {
-  if (str.length > size) {
-    return true;
-  }
-  return false;
-};
 
 /* check string is a web link */
 export function strIsLink(str?: string) {
@@ -38,135 +25,6 @@ export const simpleText = (text = '') => {
 
   return text;
 };
-
-export const valToStr = (val: any) => {
-  if (val === undefined) return '';
-  if (val === null) return 'null';
-
-  if (typeof val === 'object') {
-    try {
-      const start = Date.now();
-      const res = JSON.stringify(val);
-
-      if (Date.now() - start > 1000) {
-        console.warn('Slow JSON.stringify', {
-          duration: Date.now() - start,
-          valLength: res.length
-        });
-      }
-
-      return res;
-    } catch (error) {
-      console.error('Failed to stringify value', { error });
-      return `Failed to stringify value: ${getErrText(error)}`;
-    }
-  }
-
-  return String(val);
-};
-
-const VARIABLE_PLACEHOLDER_PATTERN = /\{\{([^}]+)\}\}/g;
-const SKIP_VARIABLE_REPLACEMENT = Symbol('skipVariableReplacement');
-type ReplaceVariableOptions = {
-  depth?: number;
-  maxStringLength?: number;
-};
-
-const hasVariableKey = (obj: Record<string, any>, key: string) => {
-  if (Object.prototype.hasOwnProperty.call(obj, key)) {
-    return Object.prototype.propertyIsEnumerable.call(obj, key);
-  }
-
-  if (!(key in obj)) return false;
-
-  let proto = Object.getPrototypeOf(obj);
-  while (proto) {
-    if (Object.prototype.hasOwnProperty.call(proto, key)) {
-      return Object.prototype.propertyIsEnumerable.call(proto, key);
-    }
-    proto = Object.getPrototypeOf(proto);
-  }
-
-  // Proxy-backed variable records can expose virtual keys through the `has` trap.
-  return true;
-};
-
-/**
- * 将文本中的 `{{variable}}` 占位符替换为变量值。
- *
- * 该函数位于 global 包，不能直接读取 service env；服务侧需要把已校验的字符串上限显式传入。
- * 替换时只格式化模板实际引用的变量，避免大变量表在每次工作流节点运行时被整体 stringify。
- */
-export function replaceVariable(
-  text: any,
-  obj: Record<string, any>,
-  optionsOrDepth: ReplaceVariableOptions | number = {},
-  fallbackMaxStringLength = DEFAULT_MAX_STRING_LENGTH
-) {
-  const { depth, maxStringLength } =
-    typeof optionsOrDepth === 'number'
-      ? { depth: optionsOrDepth, maxStringLength: fallbackMaxStringLength }
-      : {
-          depth: optionsOrDepth.depth ?? 0,
-          maxStringLength: optionsOrDepth.maxStringLength ?? DEFAULT_MAX_STRING_LENGTH
-        };
-
-  if (typeof text !== 'string') return text;
-  if (checkStrOversize(text, maxStringLength)) {
-    throw new Error(getTextOversizeErrorMessage(maxStringLength));
-  }
-  if (!text.includes('{{')) return text;
-
-  const MAX_REPLACEMENT_DEPTH = 10;
-  if (depth > MAX_REPLACEMENT_DEPTH) {
-    return text;
-  }
-
-  const hasCircularReference = (value: any, targetKey: string): boolean => {
-    if (typeof value !== 'string') return false;
-
-    const selfRefPattern = new RegExp(
-      `\\{\\{${targetKey.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')}\\}\\}`,
-      'g'
-    );
-    return selfRefPattern.test(value);
-  };
-
-  let result = text;
-  let currentDepth = depth;
-
-  while (currentDepth <= MAX_REPLACEMENT_DEPTH && result.includes('{{')) {
-    let changed = false;
-    const replacementCache = new Map<string, string | typeof SKIP_VARIABLE_REPLACEMENT>();
-
-    result = result.replace(VARIABLE_PLACEHOLDER_PATTERN, (match: string, key: string) => {
-      if (!hasVariableKey(obj, key)) return match;
-
-      if (replacementCache.has(key)) {
-        const cachedReplacement = replacementCache.get(key);
-        return cachedReplacement === SKIP_VARIABLE_REPLACEMENT ? match : (cachedReplacement ?? '');
-      }
-
-      const val = obj[key];
-      if (hasCircularReference(val, key)) {
-        replacementCache.set(key, SKIP_VARIABLE_REPLACEMENT);
-        return match;
-      }
-
-      const replacement = valToStr(val);
-      replacementCache.set(key, replacement);
-      if (replacement !== match) {
-        changed = true;
-      }
-      return replacement;
-    });
-
-    if (!changed) break;
-    currentDepth++;
-  }
-
-  return result || '';
-}
 
 /* replace sensitive text */
 export const replaceSensitiveText = (text: string) => {
