+++ packages/global/core/workflow/runtime/utils.ts
@@ -1,11 +1,4 @@
 import json5 from 'json5';
-import {
-  checkStrOversize,
-  DEFAULT_MAX_STRING_LENGTH,
-  getTextOversizeErrorMessage,
-  replaceVariable,
-  valToStr
-} from '../../../common/string/tools';
 import { ChatRoleEnum } from '../../../core/chat/constants';
 import type { ChatItemMiniType } from '../../../core/chat/type';
 import type { NodeOutputItemType } from './type';
@@ -366,136 +359,6 @@ export const formatVariableValByType = (val: any, valueType?: WorkflowIOValueTyp
 
   return val;
 };
-
-// 模块级 RegExp 缓存，避免每次变量替换都重新编译正则
-const _replaceRegexCache = new Map<string, RegExp>();
-const _MAX_REGEX_CACHE_SIZE = 5000;
-
-const _getCachedRegex = (pattern: string): RegExp => {
-  let re = _replaceRegexCache.get(pattern);
-  if (!re) {
-    if (_replaceRegexCache.size >= _MAX_REGEX_CACHE_SIZE) {
-      _replaceRegexCache.clear();
-    }
-    re = new RegExp(pattern, 'g');
-    _replaceRegexCache.set(pattern, re);
-  }
-  return re;
-};
-
-// replace {{$xx.xx$}} variables for text
-export function replaceEditorVariable({
-  text,
-  nodesMap,
-  variables,
-  depth = 0,
-  maxStringLength = DEFAULT_MAX_STRING_LENGTH
-}: {
-  text: any;
-  nodesMap: Record<string, RuntimeNodeItemType> | Map<string, RuntimeNodeItemType>;
-  variables: Record<string, unknown>; // runtime global variables
-  depth?: number;
-  maxStringLength?: number;
-}) {
-  const getNode = (nodeId: string) => {
-    return nodesMap instanceof Map ? nodesMap.get(nodeId) : nodesMap[nodeId];
-  };
-  if (typeof text !== 'string') return text;
-  if (text === '') return text;
-  if (checkStrOversize(text, maxStringLength)) {
-    throw new Error(getTextOversizeErrorMessage(maxStringLength));
-  }
-  if (!text.includes('{{')) return text;
-
-  const MAX_REPLACEMENT_DEPTH = 10;
-
-  // Prevent infinite recursion
-  if (depth > MAX_REPLACEMENT_DEPTH) {
-    return text;
-  }
-
-  text = replaceVariable(text, variables, { maxStringLength });
-
-  // Check for circular references in variable values
-  const hasCircularReference = (value: any, targetKey: string): boolean => {
-    if (typeof value !== 'string') return false;
-
-    // Check if the value contains the target variable pattern (direct self-reference)
-    const selfRefPattern = _getCachedRegex(
-      `\\{\\{\\$${targetKey.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')}\\$\\}\\}`
-    );
-    selfRefPattern.lastIndex = 0;
-    return selfRefPattern.test(value);
-  };
-
-  const variablePattern = /\{\{\$([^.]+)\.([^$]+)\$\}\}/g;
-  let result = text;
-  let currentDepth = depth;
-
-  while (currentDepth <= MAX_REPLACEMENT_DEPTH && result.includes('{{$')) {
-    let hasReplacements = false;
-    const replacementCache = new Map<string, string | undefined>();
-
-    result = result.replace(variablePattern, (match: string, nodeId: string, id: string) => {
-      const variableKey = `${nodeId}.${id}`;
-      if (replacementCache.has(variableKey)) {
-        const cachedReplacement = replacementCache.get(variableKey);
-        return cachedReplacement === undefined ? match : cachedReplacement;
-      }
-
-      const variableVal = (() => {
-        if (nodeId === VARIABLE_NODE_ID) {
-          return variables[id];
-        }
-
-        // Find upstream node input/output
-        const node = getNode(nodeId);
-        if (!node) return;
-
-        const output = node.outputs.find((output) => output.id === id);
-        if (output) return formatVariableValByType(output.value, output.valueType);
-
-        // Use the node's input as the variable value(Example: HTTP data will reference its own dynamic input)
-        const input = node.inputs.find((input) => input.key === id);
-        if (input) {
-          return getReferenceVariableValue({
-            value: input.value,
-            nodesMap,
-            variables
-          });
-        }
-      })();
-
-      // 直接自引用保持原占位符，交给最大深度保护兜底更复杂的环。
-      if (hasCircularReference(variableVal, variableKey)) {
-        replacementCache.set(variableKey, undefined);
-        return match;
-      }
-
-      const replacement = valToStr(variableVal);
-      replacementCache.set(variableKey, replacement);
-      if (replacement !== match) {
-        hasReplacements = true;
-      }
-      return replacement;
-    });
-
-    if (!hasReplacements) break;
-    currentDepth++;
-
-    if (checkStrOversize(result, maxStringLength)) {
-      console.warn(getTextOversizeErrorMessage(maxStringLength));
-      break;
-    }
-
-    // 旧逻辑每次处理嵌套节点引用前都会先处理普通变量，这里保留该顺序。
-    if (currentDepth <= MAX_REPLACEMENT_DEPTH && result.includes('{{$')) {
-      result = replaceVariable(result, variables, { maxStringLength });
-    }
-  }
-
-  return result || '';
-}
 
 export const textAdaptGptResponse = ({
   text,
