+++ packages/global/test/common/string/tools.test.ts
@@ -1,19 +1,15 @@
 import { describe, expect, it } from 'vitest';
 import {
   customNanoid,
-  DEFAULT_MAX_STRING_LENGTH,
   formatNumberWithUnit,
-  getTextOversizeErrorMessage,
   getNanoid,
   hashStr,
   replaceRegChars,
   replaceSensitiveText,
-  replaceVariable,
   simpleText,
   sliceJsonStr,
   sliceStrStartEnd,
-  strIsLink,
-  valToStr
+  strIsLink
 } from '@fastgpt/global/common/string/tools';
 
 describe('string tools', () => {
@@ -37,107 +33,6 @@ describe('string tools', () => {
     const input = '  你 好 \r\n\r\n\r\nfoo\x00bar  ';
     expect(simpleText(input)).toBe('你 好 \n\nfoo bar');
     expect(simpleText('a   b')).toBe('a   b');
-  });
-
-  it('should convert values to strings', () => {
-    expect(valToStr(undefined)).toBe('');
-    expect(valToStr(null)).toBe('null');
-    expect(valToStr({ a: 1 })).toBe('{"a":1}');
-    expect(valToStr(123)).toBe('123');
-  });
-
-  it('should replace variables with recursion and safeguards', () => {
-    expect(replaceVariable('Hello {{name}}', { name: 'Ada' })).toBe('Hello Ada');
-    expect(
-      replaceVariable('Hello {{name}}', {
-        name: '{{first}} {{last}}',
-        first: 'Ada',
-        last: 'Lovelace'
-      })
-    ).toBe('Hello Ada Lovelace');
-    expect(replaceVariable('Hello {{name}}', { name: undefined })).toBe('Hello ');
-    expect(replaceVariable('Hello {{name}}', { name: '{{name}}' })).toBe('Hello {{name}}');
-    expect(replaceVariable(123 as any, { name: 'Ada' })).toBe(123);
-  });
-
-  it('should only stringify variables that appear in the template', () => {
-    let stringifyCount = 0;
-    const unusedLargeObject = {
-      toJSON() {
-        stringifyCount += 1;
-        return { value: 'unused' };
-      }
-    };
-
-    expect(
-      replaceVariable('Hello {{name}}', {
-        name: 'Ada',
-        unusedLargeObject
-      })
-    ).toBe('Hello Ada');
-    expect(stringifyCount).toBe(0);
-  });
-
-  it('should return strings without placeholders before reading variables', () => {
-    let stringifyCount = 0;
-    const unused = {
-      toJSON() {
-        stringifyCount += 1;
-        return { value: 'unused' };
-      }
-    };
-
-    expect(replaceVariable('Hello Ada', { unused })).toBe('Hello Ada');
-    expect(stringifyCount).toBe(0);
-  });
-
-  it('should use an explicit max string length without reading env', () => {
-    expect(DEFAULT_MAX_STRING_LENGTH).toBe(100_000_000);
-    expect(getTextOversizeErrorMessage()).toBe('Text length exceeds 100,000,000 characters.');
-    expect(getTextOversizeErrorMessage(2_000_000)).toBe(
-      'Text length exceeds 2,000,000 characters.'
-    );
-
-    expect(() =>
-      replaceVariable('hello {{name}}', { name: 'Ada' }, { maxStringLength: 5 })
-    ).toThrow('Text length exceeds 5 characters.');
-  });
-
-  it('should stringify the same referenced variable once per replacement round', () => {
-    let stringifyCount = 0;
-    const value = {
-      toJSON() {
-        stringifyCount += 1;
-        return { a: 1 };
-      }
-    };
-
-    expect(replaceVariable('{{value}} {{value}}', { value })).toBe('{"a":1} {"a":1}');
-    expect(stringifyCount).toBe(1);
-  });
-
-  it('should not replace non-enumerable Object prototype keys', () => {
-    expect(replaceVariable('value: {{toString}}', {})).toBe('value: {{toString}}');
-    expect(replaceVariable('value: {{toString}}', { toString: 'own value' })).toBe(
-      'value: own value'
-    );
-  });
-
-  it('should treat $ special characters in replacement value as literals', () => {
-    // $1, $2 不应被解释为捕获组引用
-    expect(replaceVariable('value: {{val}}', { val: '$1' })).toBe('value: $1');
-    expect(replaceVariable('value: {{val}}', { val: '$2' })).toBe('value: $2');
-    // $$ 不应被解释为字面量 $
-    expect(replaceVariable('value: {{val}}', { val: '$$' })).toBe('value: $$');
-    // $& 不应被替换为整个匹配字符串
-    expect(replaceVariable('value: {{val}}', { val: '$&' })).toBe('value: $&');
-    // $` 和 $' 不应被替换为匹配前/后的内容
-    expect(replaceVariable('value: {{val}}', { val: "$'" })).toBe("value: $'");
-    expect(replaceVariable('value: {{val}}', { val: '$`' })).toBe('value: $`');
-    // 混合场景
-    expect(replaceVariable('result={{a}}&other={{b}}', { a: '$1', b: '$2' })).toBe(
-      'result=$1&other=$2'
-    );
   });
 
   it('should replace sensitive text', () => {
