+++ packages/service/core/ai/llm/promptCall/prompt.ts
@@ -1,6 +1,5 @@
-import { replaceVariable } from '@fastgpt/global/common/string/tools';
+import { replaceVariable } from '../../../../common/string/replaceVariable';
 import type { ChatCompletionTool } from '@fastgpt/global/core/ai/llm/type';
-import { SYSTEM_MAX_STRING_LENGTH } from '../../../../env';
 
 export const getPromptToolCallPrompt = (tools: ChatCompletionTool['function'][]) => {
   const prompt = `<ToolSkill>
@@ -36,13 +35,7 @@ export const getPromptToolCallPrompt = (tools: ChatCompletionTool['function'][])
     parameters: tool.parameters
   }));
 
-  return replaceVariable(
-    prompt,
-    {
-      toolSchema: JSON.stringify(schema)
-    },
-    {
-      maxStringLength: SYSTEM_MAX_STRING_LENGTH
-    }
-  );
+  return replaceVariable(prompt, {
+    toolSchema: JSON.stringify(schema)
+  });
 };
