+++ projects/app/src/service/core/dataset/queues/generateQA.ts
@@ -3,7 +3,7 @@ import { pushLLMTrainingUsage } from '@fastgpt/service/support/wallet/usage/cont
 import { TrainingModeEnum } from '@fastgpt/global/core/dataset/constants';
 import type { ChatCompletionMessageParam } from '@fastgpt/global/core/ai/llm/type';
 import { getLogger, LogCategories } from '@fastgpt/service/common/logger';
-import { replaceVariable } from '@fastgpt/global/common/string/tools';
+import { replaceVariable } from '@fastgpt/service/common/string/replaceVariable';
 import { Prompt_AgentQA } from '@fastgpt/global/core/ai/prompt/agent';
 import type { PushDataChunkType } from '@fastgpt/global/openapi/core/dataset/data/api';
 import { getLLMModel } from '@fastgpt/service/core/ai/model';
@@ -20,7 +20,6 @@ import { pushDataListToTrainingQueue } from '@fastgpt/service/core/dataset/train
 import { delay } from '@fastgpt/service/common/bullmq';
 import { createLLMResponse } from '@fastgpt/service/core/ai/llm/request';
 import { UsageItemTypeEnum } from '@fastgpt/global/support/wallet/usage/constants';
-import { SYSTEM_MAX_STRING_LENGTH } from '@fastgpt/service/env';
 
 const logger = getLogger(LogCategories.MODULE.DATASET.QA);
 
@@ -128,7 +127,7 @@ export async function generateQA(): Promise<any> {
       try {
         const modelData = getLLMModel(data.dataset.agentModel);
         const prompt = `${data.collection.qaPrompt || Prompt_AgentQA.description}
-  ${replaceVariable(Prompt_AgentQA.fixedText, { text }, { maxStringLength: SYSTEM_MAX_STRING_LENGTH })}`;
+  ${replaceVariable(Prompt_AgentQA.fixedText, { text })}`;
 
         // request LLM to get QA
         const messages: ChatCompletionMessageParam[] = [
